Make colors RGB again
The hardware never considered the lights to be in GRB order. However,
there was something else going on. Basically, the library is written to
always allow you to specify colors in RGB order, but it can then write
those colors to the lights in different orders (mainly RGB or GRB) and
for some reason, by default it assumes your LEDs take the colors in GRB
order.

As such, it was rewriting the RGB colors to GRB format, even though the
LEDs actually took them in RGB order, which made the colors wrong.
Writing the colors as GRB is the opposite of this, so it undoes the
change, making it seem that's the right way to specify the colors.

By just telling the library that the LEDs are taking the colors in RGB
order, you can specify the colors the way you normally do: RGB.